requirements.txt
Django==5.2.6
pillow==11.3.0
crispy-tailwind==1.0.3
djangorestframework==3.16.1
django-environ==0.12.0
psycopg2-binary==2.9.10